event-analytics
events belonging to an organizer will only be displayed in the organizer dashboard and the csv can be exported. I will work on the additional details of specifically exporting the attendees for an event, but I'm thinking it might be part of sprint 3 cause it has to deal with students and organizer roles. They're connected so it can't really work at the moment since we have no student functionalities set up.

users with admin role will only see admin dashboard.
when creating a new account, the default role will be student and they have no access to dashboard.