admin management - organizer accounts mod
This is part of the admin features under platform oversight and management. the admin can now see which student account made a request to become an organizer and for which event. Changes that were made also reflect in organizer dashboard. When a student becomes an organizer for a certain event(s), they now have access to the dashboard for those events in particular.