Bug related to issue #167
If the users created an account with a single interest, it would cause a error which would then display their username as 'error'. This is caused by the fact that a single interest selected was stored as a string and not an array while the code could only handle arrays. It has been changed so that it can handle both and to be EXTRA SURE, a migration script has been created.